backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.125.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
SQLAlchemy==2.0.36
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
